improve download response

+++ src/LaravelRecordProcessor/Stages/DownloadFileResponse.php
@@ -14,10 +14,14 @@ class DownloadFileResponse extends DownloadFileOutput implements ProcessorStageF
             throw new InvalidArgumentException( 'Cannot use a temporary file to make a download response' );
         }
 
+        if (!$this->inputFileInfo->isReadable() || $this->inputFileInfo->getRealPath() === false) {
+            throw new InvalidArgumentException( 'Cannot read file for download' );
+        }
+
         $filename = rawurlencode( $this->outputFileInfo->getBasename() );
 
         return response()
-            ->download( $this->inputFile, $filename )
+            ->download( $this->inputFileInfo, $filename )
             ->deleteFileAfterSend( $this->deleteAfterDownload );
     }
 }
